Data.TASequence.ToCatQueue

Description

A purely functional catenable queue representation with that turns takes a purely functional queue and turns in it into a catenable queue, i.e. with the same complexity for >< as for |> Based on Purely functional data structures by Chris Okasaki section 7.2: Catenable lists

Documentation

module Data.TASequence

data ToCatQueue q c x y

The catenable queue type. The first type argument is the type of the queue we use (|>)